requirements.txt
pandas
numpy
yfinance
statsmodels
scikit-learn
hmmlearn
arch
pyyaml
matplotlib
seaborn
kiteconnect
jupyter
